[update] アカウント作成、創作者存在確認のテストを実装
